Converts VideoThumnail to react effects for redux
