+++ TTKDTrainingApp/src/components/Content.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {View, Text, Button, ScrollView, StyleSheet} from 'react-native';
 
 import VideoWithControlsConnector from './VideoWithControls';
-import VideoThumbnailConnector from './VideoThumbnail';
+import VideoThumbnail from './VideoThumbnail';
 
 const Content = props => {
   const {navigation} = props;
@@ -36,10 +36,7 @@ const Content = props => {
       <VideoWithControlsConnector contentVideo={content.video} />
       {recordButton}
       <ScrollView horizontal={true}>
-        <VideoThumbnailConnector
-          navigation={navigation}
-          contentId={content.id}
-        />
+        <VideoThumbnail navigation={navigation} contentId={content.id} />
       </ScrollView>
     </View>
   );
