--- TTKDTrainingApp/src/components/VideoThumbnail/index.js
@@ -1,3 +1,0 @@
-import {VideoThumbnailConnector} from './VideoThumbnailConnector';
-
-export default VideoThumbnailConnector;
